+++ libs/orm-designer/src/postgres/reference/reference-resource/reference-resource.orm.ts
@@ -1,3 +1,5 @@
+import { ReferenceResourceType } from '@voiceflow/dtos';
+
 import { PostgresObjectIDMutableORM } from '@/postgres/common/orms/postgres-object-id-mutable.orm';
 
 import { ReferenceResourceEntity } from './reference-resource.entity';
@@ -14,5 +16,16 @@ export class ReferenceResourceORM extends PostgresObjectIDMutableORM<ReferenceRe
 
   deleteManyByEnvironment(environmentID: string) {
     return this.delete({ environmentID });
+  }
+
+  async findOneByTypeDiagramIDAndResourceID(data: {
+    type: ReferenceResourceType;
+    diagramID: string | null;
+    resourceID: string;
+    environmentID: string;
+  }) {
+    const [resource = null] = await this.find(data, { limit: 1 });
+
+    return resource;
   }
 }